requirements.txt
PyYAML
python-box
ensure

-e .